Adjust data type of capacity and channel. -1 channel means the character is offline.

+++ atlas.com/buddies/kafka/consumer/list/kafka.go
@@ -15,7 +15,7 @@ type command[E any] struct {
 }
 
 type createCommandBody struct {
-	Capacity uint32 `json:"capacity"`
+	Capacity byte `json:"capacity"`
 }
 
 type requestAddBuddyCommandBody struct {
